+++ src/Fields/DatePicker.php
@@ -17,6 +17,7 @@ use WordPlate\Acf\Fields\Attributes\ConditionalLogic;
 use WordPlate\Acf\Fields\Attributes\DateTimeFormat;
 use WordPlate\Acf\Fields\Attributes\Disabled;
 use WordPlate\Acf\Fields\Attributes\Instructions;
+use WordPlate\Acf\Fields\Attributes\ReadOnly;
 use WordPlate\Acf\Fields\Attributes\Required;
 use WordPlate\Acf\Fields\Attributes\WeekDay;
 use WordPlate\Acf\Fields\Attributes\Wrapper;
@@ -27,6 +28,7 @@ class DatePicker extends Field
     use DateTimeFormat;
     use Disabled;
     use Instructions;
+    use ReadOnly;
     use Required;
     use WeekDay;
     use Wrapper;