+++ src/Fields/DateTimePicker.php
@@ -15,7 +15,9 @@ namespace WordPlate\Acf\Fields;
 
 use WordPlate\Acf\Fields\Attributes\ConditionalLogic;
 use WordPlate\Acf\Fields\Attributes\DateTimeFormat;
+use WordPlate\Acf\Fields\Attributes\Disabled;
 use WordPlate\Acf\Fields\Attributes\Instructions;
+use WordPlate\Acf\Fields\Attributes\ReadOnly;
 use WordPlate\Acf\Fields\Attributes\Required;
 use WordPlate\Acf\Fields\Attributes\WeekDay;
 use WordPlate\Acf\Fields\Attributes\Wrapper;
@@ -24,7 +26,9 @@ class DateTimePicker extends Field
 {
     use ConditionalLogic;
     use DateTimeFormat;
+    use Disabled;
     use Instructions;
+    use ReadOnly;
     use Required;
     use WeekDay;
     use Wrapper;